requirements.txt
fastapi
uvicorn
langchain-groq
langgraph
langchain-core
pypdf2
python-multipart
python-dotenv
pydantic
pydantic-settings
sqlalchemy
asyncpg
alembic
testcontainers
pytest
pytest-asyncio
httpx
ruff
mypy
pre-commit
